split generation tests into own file
