+++ test/index.js
@@ -2,3 +2,4 @@ require('./grid_test.js');
 require('./cell_test.js');
 require('./room_test.js');
 require('./display_handler_test.js');
+require('./generator_test.js');
